requirements.txt
streamlit
pandas
Pillow
python-dateutil